Adds order statistics by country
Implements endpoint to retrieve order statistics grouped by country.
This allows for analysis of order distribution across different regions.

The frontend is updated to use the new endpoint and pass the id of the customer account, fixing a data fetching issue on the client dashboard.
Also added console.log for debug.